refactor: Adaptor 추상화를 적용한다.
