+++ mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -4,18 +4,15 @@ import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.controller.tobe.HandlerExecution;
-import nextstep.mvc.controller.tobe.exception.FailMapHandler;
-import nextstep.mvc.view.JspView;
+import nextstep.mvc.exception.FailMapHandler;
+import nextstep.mvc.handlerAdaptor.HandlerAdapter;
+import nextstep.mvc.handlerMapping.HandlerMapping;
 import nextstep.mvc.view.ModelAndView;
-import nextstep.mvc.view.View;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 
 public class DispatcherServlet extends HttpServlet {
@@ -24,9 +21,11 @@ public class DispatcherServlet extends HttpServlet {
     private static final Logger log = LoggerFactory.getLogger(DispatcherServlet.class);
 
     private final List<HandlerMapping> handlerMappings;
+    private final List<HandlerAdapter> handlerAdapters;
 
     public DispatcherServlet() {
         this.handlerMappings = new ArrayList<>();
+        this.handlerAdapters = new ArrayList<>();
     }
 
     @Override
@@ -38,15 +37,22 @@ public class DispatcherServlet extends HttpServlet {
         handlerMappings.add(handlerMapping);
     }
 
+    public void addHandlerAdapter(final HandlerAdapter handlerAdapter) {
+        handlerAdapters.add(handlerAdapter);
+    }
+
     @Override
     protected void service(final HttpServletRequest request, final HttpServletResponse response) throws ServletException {
         log.debug("Method : {}, Request URI : {}", request.getMethod(), request.getRequestURI());
-        final Object handler = getHandler(request);
-        if (handler instanceof Controller) {
-            handleManualHandler(request, response, (Controller) handler);
-        }
-        if (handler instanceof HandlerExecution) {
-            handleAnnotationHandler(request, response, (HandlerExecution) handler);
+
+        try {
+            final Object handler = getHandler(request);
+            final HandlerAdapter handlerAdapter = getHandlerAdapter(handler);
+            final ModelAndView modelAndView = handlerAdapter.handle(request, response, handler);
+            modelAndView.render(request, response);
+        } catch (Throwable e) {
+            log.error("Exception : {}", e.getMessage(), e);
+            throw new ServletException(e.getMessage());
         }
     }
 
@@ -58,33 +64,10 @@ public class DispatcherServlet extends HttpServlet {
                 .orElseThrow(FailMapHandler::new);
     }
 
-    private void handleManualHandler(final HttpServletRequest request, final HttpServletResponse response,
-                                     final Controller handler) throws ServletException {
-        try {
-            final String viewName = handler.execute(request, response);
-            final ModelAndView modelAndView = new ModelAndView(new JspView(viewName));
-            renderView(modelAndView, request, response);
-        } catch (Throwable e) {
-            log.error("Exception : {}", e.getMessage(), e);
-            throw new ServletException(e.getMessage());
-        }
-    }
-
-    private void handleAnnotationHandler(final HttpServletRequest request, final HttpServletResponse response,
-                                         final HandlerExecution handler) {
-        try {
-            final ModelAndView modelAndView = handler.handle(request, response);
-            renderView(modelAndView, request, response);
-        } catch (Exception e) {
-            log.error("Exception : {}", e.getMessage(), e);
-            throw new RuntimeException(e.getMessage());
-        }
-    }
-
-    private static void renderView(final ModelAndView modelAndView, final HttpServletRequest request,
-                                   final HttpServletResponse response) throws Exception {
-        final Map<String, Object> model = modelAndView.getModel();
-        final View view = modelAndView.getView();
-        view.render(model, request, response);
+    private HandlerAdapter getHandlerAdapter(final Object handler) {
+        return handlerAdapters.stream()
+                .filter(handlerAdapter -> handlerAdapter.supports(handler))
+                .findFirst()
+                .orElseThrow(FailMapHandler::new);
     }
 }
